mcp/requirements.txt
# MCP Server Requirements for LocalTaskClawXiaozhi

# MCP framework
fastmcp>=2.13.0
mcp>=1.20.0

# HTTP client for Xiaozhi API
httpx>=0.27.0

# Environment variables
python-dotenv>=1.2.1

# WebSocket support
websockets>=15.0.1

# Proxy for remote MCP servers (optional)
mcp-proxy>=0.10.0
